+++ src/Sentry/SentryOptions.cs
@@ -31,7 +31,7 @@ namespace Sentry;
 #if __MOBILE__
 public partial class SentryOptions
 #else
-public partial class SentryOptions
+public class SentryOptions
 #endif
 {
     private Dictionary<string, string>? _defaultTags;
@@ -1370,6 +1370,8 @@ public partial class SentryOptions
         );
 
         NetworkStatusListener = new PollingNetworkStatusListener(this);
+
+        Experimental = new ExperimentalSentryOptions(this);
     }
 
     /// <summary>
@@ -1912,7 +1914,7 @@ public partial class SentryOptions
     /// <remarks>
     /// Experimental features are subject to binary, source and behavioral breaking changes in future updates.
     /// </remarks>
-    public ExperimentalSentryOptions Experimental { get; } = new();
+    public ExperimentalSentryOptions Experimental { get; }
 
     /// <summary>
     /// Sentry features that are currently in an experimental state.
@@ -1922,13 +1924,15 @@ public partial class SentryOptions
     /// </remarks>
     public class ExperimentalSentryOptions
     {
-        private TraceMetricsCallbacks? _beforeSendMetric;
+        private readonly SentryOptions _options;
+        private SentryTraceMetricsCallbacks? _beforeSendMetric;
 
-        internal ExperimentalSentryOptions()
+        internal ExperimentalSentryOptions(SentryOptions options)
         {
+            _options = options;
         }
 
-        internal TraceMetricsCallbacks? BeforeSendMetricInternal => _beforeSendMetric;
+        internal SentryTraceMetricsCallbacks? BeforeSendMetricInternal => _beforeSendMetric;
 
         /// <summary>
         /// When set to <see langword="false"/>, the SDK does not generate and send metrics to Sentry via <see cref="SentrySdk"/>.
@@ -1938,18 +1942,20 @@ public partial class SentryOptions
         public bool EnableMetrics { get; set; } = true;
 
         /// <summary>
-        /// Sets a callback function to be invoked before sending the metric to Sentry.
+        /// Sets a callback function to be invoked before sending the metric of numeric type <typeparamref name="T"/> to Sentry.
         /// When the delegate throws an <see cref="Exception"/> during invocation, the metric will not be captured.
         /// </summary>
+        /// <typeparam name="T">The numeric type of the metric.</typeparam>
         /// <remarks>
         /// It can be used to modify the metric object before being sent to Sentry.
         /// To prevent the metric from being sent to Sentry, return <see langword="null"/>.
+        /// Supported numeric value types for <typeparamref name="T"/> are <see langword="byte"/>, <see langword="short"/>, <see langword="int"/>, <see langword="long"/>, <see langword="float"/>, and <see langword="double"/>.
         /// </remarks>
         /// <seealso href="https://develop.sentry.dev/sdk/telemetry/metrics/"/>
         public void SetBeforeSendMetric<T>(Func<SentryMetric<T>, SentryMetric<T>?> beforeSendMetric) where T : struct
         {
-            _beforeSendMetric ??= new TraceMetricsCallbacks();
-            _beforeSendMetric.Set(beforeSendMetric);
+            _beforeSendMetric ??= new SentryTraceMetricsCallbacks();
+            _beforeSendMetric.Set(beforeSendMetric, _options.DiagnosticLogger);
         }
     }
 }